gator-driver: Hack makefile to search for Mali trace headers
Gator needs Mali headers and device type in order to build with Mali
profiling support. Unfortunately, the presence and location of this
information is difficult to configure considering kernels are
separate packages and managed by various teams.

To work around this, just use the find command to look for the header
and if present, assume we want Mali support, and that it is a Mali-400.

Signed-off-by: Jon Medhurst <tixy@linaro.org>