gator: Prevent BUG() when no device-tree cpu nodes present.
When IKS support is enabled in gator but we are running on boards
without a device-tree or where there are no cpu nodes in the
device-tree, then calc_first_cluster_size will call BUG_ON() because
mpidr_cpuids_count == 0.

To work around this, we will instead set a flag to indicate we haven't
managed to create an mpidr table and fallback to the behaviour we would
have if IKS wasn't enabled. This means that IKS support will only
function as expected if there are device-tree nodes for CPUs but we
expect that to always be the case anyway.

Signed-off-by: Jon Medhurst <tixy@linaro.org>